# kv: return pointers from TxnSender.GetLeafTxn{Input/Final}State