--- a/pkg/sql/distsql_running.go
+++ b/pkg/sql/distsql_running.go
@@ -456,7 +456,7 @@ func (dsp *DistSQLPlanner) Run(
 			recv.SetError(err)
 			return func() {}
 		}
-		leafInputState = &tis
+		leafInputState = tis
 	}
 
 	if logPlanDiagram {